Even more internal changes
