+++ sella/__init__.py
@@ -1,4 +1,3 @@
-from .peswrapper import PESWrapper
 from .optimize import optimize
 from .irc import IRC
 from .aseopt import Sella
